Fixing manipulator values

+++ src/robot_module/src/robot_module/services/manipulator.py
@@ -7,7 +7,7 @@ import rospy
 import threading
 from .service import Service
 
-from std_msgs.msg import Int32
+from std_msgs.msg import Float32
 
 
 topic = "/nautilus/manipulator/pwm"
@@ -20,9 +20,9 @@ class Manipulator(Service):
     """
     def __init__(self, node_name):
         super().__init__(node_name)
-        self.msg = Int32()
+        self.msg = Float32()
 
-        self.mani_pub = rospy.Publisher(topic, Int32, queue_size=10)
+        self.mani_pub = rospy.Publisher(topic, Float32, queue_size=10)
 
 
     def setup(self):
